fetching latest from main

+++ azure_arc_k8s_jumpstart/aks_hybrid/aks_edge_essentials_full_akri/bicep_template/artifacts/LogonScript.ps1
@@ -14,12 +14,13 @@ $azureLocation = $env:azureLocation
 $spnClientId = $env:spnClientId
 $spnClientSecret = $env:spnClientSecret
 $spnTenantId = $env:spnTenantId
-$vhdxUri = "https://jsvhds.blob.core.windows.net/scenarios/prod/JSW11IoTBase.vhdx?sp=r&st=2023-05-09T12:36:32Z&se=2033-05-09T20:36:32Z&spr=https&sv=2022-11-02&sr=b&sig=xFROrqGkKDIdrXqiAMLZGLEwTMToOWoNNDMVz1zvPMc%3D"
+$vhdxUri = "https://jumpstartprodsg.blob.core.windows.net/scenarios/prod/JSW11IoTBase.vhdx"
 $hypervVMUser = "Administrator"
 $hypervVMPassword = "JS123!!"
 $kubernetesDistribution = $env:kubernetesDistribution
 $templateBaseUrl = $env:templateBaseUrl
 $aksEEReleasesUrl = "https://api.github.com/repos/Azure/AKS-Edge/releases"
+$AKSEEPinnedSchemaVersion = $Env:AKSEEPinnedSchemaVersion
 
 Write-Header "Executing LogonScript.ps1"
 
@@ -233,7 +234,6 @@ Invoke-Command -VMName $VMnames -Credential $Credentials -ScriptBlock {
     $deploymentFolder = "C:\Deployment" # Deployment folder is already pre-created in the VHD image
     $logsFolder = "$deploymentFolder\Logs"
     $kubeFolder = "$env:USERPROFILE\.kube"
-
     # Set up an array of folders
     $folders = @($logsFolder, $kubeFolder)
 
@@ -252,6 +252,7 @@ Invoke-Command -VMName $VMnames -Credential $Credentials -ScriptBlock {
     $logsFolder = "$deploymentFolder\Logs"
     Start-Transcript -Path $logsFolder\AKSEEBootstrap.log
     $SiteConfig = $using:SiteConfig
+    $AKSEEPinnedSchemaVersion = $using:AKSEEPinnedSchemaVersion
 
     ##########################################
     # Deploying AKS Edge Essentials clusters 
@@ -303,19 +304,21 @@ Invoke-Command -VMName $VMnames -Credential $Credentials -ScriptBlock {
 }
 
 Write-Host "Fetching the latest AKS Edge Essentials release."
-$latestReleaseTag = (Invoke-WebRequest $aksEEReleasesUrl | ConvertFrom-Json)[0].tag_name
-
-$AKSEEReleaseDownloadUrl = "https://github.com/Azure/AKS-Edge/archive/refs/tags/$latestReleaseTag.zip"
-$output = Join-Path "C:\temp" "$latestReleaseTag.zip"
-Invoke-WebRequest $AKSEEReleaseDownloadUrl -OutFile $output
-Expand-Archive $output -DestinationPath "C:\temp" -Force
-$AKSEEReleaseConfigFilePath = "C:\temp\AKS-Edge-$latestReleaseTag\tools\aksedge-config.json"
-$jsonContent = Get-Content -Raw -Path $AKSEEReleaseConfigFilePath | ConvertFrom-Json
-$schemaVersion = $jsonContent.SchemaVersion
-# Clean up the downloaded release files
-Remove-Item -Path $output -Force
-Remove-Item -Path "C:\temp\AKS-Edge-$latestReleaseTag" -Force -Recurse
-
+if ($AKSEEPinnedSchemaVersion -ne "useLatest") {
+    $SchemaVersion = $AKSEEPinnedSchemaVersion
+}else{
+    $latestReleaseTag = (Invoke-WebRequest $aksEEReleasesUrl | ConvertFrom-Json)[0].tag_name
+    $AKSEEReleaseDownloadUrl = "https://github.com/Azure/AKS-Edge/archive/refs/tags/$latestReleaseTag.zip"
+    $output = Join-Path "C:\temp" "$latestReleaseTag.zip"
+    Invoke-WebRequest $AKSEEReleaseDownloadUrl -OutFile $output
+    Expand-Archive $output -DestinationPath "C:\temp" -Force
+    $AKSEEReleaseConfigFilePath = "C:\temp\AKS-Edge-$latestReleaseTag\tools\aksedge-config.json"
+    $jsonContent = Get-Content -Raw -Path $AKSEEReleaseConfigFilePath | ConvertFrom-Json
+    $schemaVersion = $jsonContent.SchemaVersion
+    # Clean up the downloaded release files
+    Remove-Item -Path $output -Force
+    Remove-Item -Path "C:\temp\AKS-Edge-$latestReleaseTag" -Force -Recurse
+}
 
 ###############################################################################
 # Setting up replacment parameters for AKS Edge Essentials config json file
